requirements.txt
atomicwrites>=1.3.0
attrs>=19.1.0
bitstring>=3.1.6
certifi>=2019.6.16
chardet>=3.0.4
click>=6.7
exifread>=2.1.2
future>=0.17.1
geopy>=1.20.0
idna>=2.8
importlib-metadata>=0.18
keyring>=8.7
keyrings.alt>=1.3
mock>=3.0.5
more-itertools>=7.1.0
packaging>=19.0
Pillow>=6.0.0
pluggy>=0.12.0
py>=1.8.0
pyicloud>=0.9.1
pyparsing>=2.4.0
PyQt5>=5.13.0
PyQt5-sip>=4.19.18
pytest>=5.0.1
pytest-qt>=3.2.2
pytz>=2019.1
PyYAML>=5.1.2
requests>=2.22.0
six>=1.12.0
tzlocal>=1.5.1
urllib3>=1.25.3
wcwidth>=0.1.7
zipp>=0.5.2
